pyproject.toml
[project]
name = "llm-eval-notes"
version = "0.1.0"
description = "LLM evaluation artifacts: hallucination, brittleness, structured output tests"
readme = "README.md"
authors = [
    { name = "Max P", email = "max.petrusenko@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "typer>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
llm-eval = "llm_eval.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
